Move displayCreators helper into a global helper file as it's used globally
